+++ src/MeisterProPR.Domain/Events/ThreadResolvedDomainEvent.cs
@@ -13,7 +13,7 @@ public sealed record ThreadResolvedDomainEvent(
     Guid ClientId,
     string RepositoryId,
     int PullRequestId,
-    int ThreadId,
+    long ThreadId,
     string? FilePath,
     string? ChangeExcerpt,
     string CommentHistory,
